[infra] Abstract Terraform configuration
Backend configuration is abstracted through `terraform init`'s
`-backend-config` flag as usual.
The block itself is commented out as a way of documenting the
convention.

Provider configuration is abstracted through the runtime environment
variables.
Their corresponding blocks (AWS and Cloudflare) are commented out as a
way of documenting the convention.

For the non-sensitive configuration values, a `variables.tf` file is
added to the setup.