[infra] Add Terraform exec scripts
A script called `exec` is implemented to have an interface to execute
Terraform commands for the application.

This is essentially the entrypoint of the IaC. It is added to discourage
executing TF commands manually from the command line and from 3rd party
vendor dependencies (e.g custom Github Actions workflows).