[infra] Use default AWS_PROFILE for CD pipeline
The CD pipeline authorizes with OIDC, which does not have the concept of
a "user profile".

"configure-aws-credentials" exports AWS_ACCESS_KEY_ID,
AWS_SECRET_ACCESS_KEY, and AWS_SESSION_TOKEN.

Therefore, the helper exec script is updated to use the default value for
AWS_PROFILE if it does not exist, which will make AWS to go through its
credential chain and pick up the required access key and secret access key.

The existance of AWS_PROFILE means that the script is executed locally.
Therefore a small comment is added to Terraform AWS provider to notify
the user.