[infra] Fix AWS_ACCOUNT_ID lookup
Apparently, the credential lookup for `aws sts` call is different than
several other AWS CLI commands and AWS SDK functions.

If a profile is passed to the command, it only checks the related
profile, it does not go through the environment variables if the profile
does not exist.

Therefore, `aws sts` call is updated to include AWS_PROFILE if it exists
(locally).