[infra] Move redirect-worker artifact generation under package.json
Since `wrangler` is not installed globally (the recommended way by
Cloudflare), the binary path must be manually entered.

The binary path depends on the installed version of `wrangler`, which is
specified under the Worker's `package.json`.

Therefore, the artifact generation command is moved under `package.json`
to easily see how `wrangler` is used, so that in the future if binary
path changes then only the Worker itself will be changed.