[infra] Cleanup the frontend artifact
Unfortunately, there is not a default way to remove the generated
artifacts in a workflow run.

Also, with `actions/upload-artifact@v4` it is not possible to upload an
artifact with the same name.

Therefore, a cleanup step is added to query the Github API and remove
the artifact manually.

Also, the artifact path in the main IaC script is updated after the
debugging session.