[infra] Add public access block resource for S3 bucket
The public access block for the S3 bucket is turned off.

This may sound anti-practice, but the access is controlled with the
policy which only allows valid Cloudflare IPs to get the objects.

Having the public access block turned on also prevents other resources
(e.g Cloudflare's Cloud Connector) to see the bucket, therefore fail
during the provision step.