[infra] Upload Cloudflare Worker as ES module
By default, the workers_script resources uploads CommonJS scripts, which
prevents us from using import statements.

Therefore, the module flag is used to support ES modules.

Also, the compatibility_flags of the Worker are updated to not include
the polyfills that are not necessary.
The size of the worker goes down from 13KB to 0.25KB with this one line
change.